// src/lib/api/types.gen.ts
export type CalendarItemDtov1 = {
  /**
   * Due Datetime
   */
  due_datetime: string;
  payment_type: PaymentType;
  /**
   * Title
   */
  title: string;
  /**
   * Amount
   */
  amount: number;
  status_type: PaymentOccurrenceStatusType | null;
};
// ...
export type PaymentOccurrenceStatusType = "PAID" | "UNPAID";

/**
 * PaymentRecurrenceType
 */
export type PaymentRecurrenceType =
  | "ONE_TIME"
  | "WEEKLY"
  | "MONTHLY"
  | "CUSTOM";

/**
 * PaymentType
 */
export type PaymentType = "EXPENSE" | "INCOME";